Move App and Logo to component folders
